+++ magda-content-api/src/Database.ts
@@ -7,6 +7,7 @@ import { Content } from "./model";
 export interface DatabaseOptions {
     dbHost: string;
     dbPort: number;
+    dbName: string;
 }
 
 export default class Database {
@@ -19,6 +20,22 @@ export default class Database {
     getContentById(id: string): Promise<Maybe<Content>> {
         return this.pool
             .query('SELECT * FROM content WHERE "id" = $1', [id])
+            .then(res => arrayToMaybe(res.rows));
+    }
+
+    setContentById(
+        id: string,
+        type: string,
+        content: string
+    ): Promise<Maybe<Content>> {
+        return this.pool
+            .query(
+                `INSERT INTO content (id, "type", "content")
+                 VALUES ($1, $2, $3)
+                 ON CONFLICT (id) DO
+                 UPDATE SET "type" = $2, "content"=$3`,
+                [id, type, content]
+            )
             .then(res => arrayToMaybe(res.rows));
     }
 }