+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 -   Document public portions of authorization api
 -   Document search api
+-   Add ability to change content (logo).
 -   Make contents API and contents database migrator for storing items which will be dynamically configurable in the future.
 -   Hyperlink organisation url on organisation page
 -   Align format facet based on its position on the page